[cgi] connector returns different values
EINCOMPLET return means to way the ESUCCESS before send data. But in
this case, the content may increase a lot before the send.
ECONTINUE return means to send data and to call again the connector. But
in this case, the script may not complete the header and it is
impossible to know the content length.