[auth] allow to set different auth rules
auth module needs the type of authentication ("Basic" or "Digest").
After the rule to check the password is an authn_<type>_<name>
sublibrary (just a C file).

Currently, only authn_basic_conf is available.

With this solution each server may has its own authentication type.
After the checking of the password is done by a library linked to the
mod_auth library.