[auth] change the euid only with UNIX shadow setting
If the user exists in the system, it is possible to use his ID, even if
the password comes from database.
If the server uses a database, and an user creates a "root" user inside
the database, he is allowed to have the root's rights. It is clearly
very dangerous.
Only user with the password of the shadow system should have the access