[userfiler] new module
After authentication and before document delivery, userfilter checks if
the user is allowed to use the URL with the method.
There is a check on the method and on the URI.
The module uses a database to store the rules. Only one rule is
available when the database is created:
 <superuser> POST <configuri>
<superuser> and <configuri> are defined in the module configuration and must
not be changed after the database creation. A special user is
"annonymous", when the connection is not authenticated.

The rules are appended in the database with a POST request on
<configuri>. The data must be
role=<user or group name>
method=<GET, POST, PUT, DELETE...>
pathexp=<regexp>

The regexp is a string to find inside the URI. There is some special
charcters:
 ^ : means the beginning of the URI, and no character must be present
before.
 $ : means the end of the URI, and no character must be present after.
 %u : is replaced by the user name.
 %g : is replaced by the user's group name.
 %h : is replaced by the user's home path.

Examples:

 role=foo&method=GET&pathexp=cgi-bin/*.cgi

The user "foo" or the users of the group "foo" is allowed to get all files
with *cgi-bin/*.cgi. barcgi-bin/index.cgi is allowed,
root/cgi-bin/passwd.cgi is allowed, cgi-bin/root/remove.cgi is allowed.

 role=users&method=PUT&pathexp=^%g/%u/*

All user of the group "users" is allowed to put file on /users/<user>/.
For "foo" /users/foo/test.txt is allowed, /users/test/test.txt is
disallowed, /root/root/test.txt is disallowed,
/cgi-bin/users/foo/text.cgi is disallowed.